Extract BearerTokenProvider into a new REST SPI module
