+++ modules/apps/oauth2-provider/oauth2-provider-jsonws/build.gradle
@@ -12,4 +12,5 @@ dependencies {
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-api")
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-scopes-liferay-api")
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-scopes-spi")
+	compileOnly project(":apps:oauth2-provider:oauth2-provider-rest-spi")
 }
