+++ modules/apps/oauth2-provider/oauth2-provider-rest-spi/bnd.bnd
@@ -1,0 +1,5 @@
+Bundle-Name: Liferay OAuth2 Provider REST SPI
+Bundle-SymbolicName: com.liferay.oauth2.provider.rest.spi
+Bundle-Version: 1.0.0
+Export-Package:\
+	com.liferay.oauth2.provider.rest.spi
