+++ modules/apps/oauth2-provider/oauth2-provider-rest-spi/build.gradle
@@ -1,0 +1,6 @@
+sourceCompatibility = "1.8"
+targetCompatibility = "1.8"
+
+dependencies {
+	compileOnly project(":apps:oauth2-provider:oauth2-provider-api")
+}
