+++ modules/apps/oauth2-provider/oauth2-provider-rest/src/main/java/com/liferay/oauth2/provider/rest/LiferayOAuthDataProvider.java
@@ -25,10 +25,10 @@ import com.liferay.oauth2.provider.model.LiferayOAuth2Scope;
 import com.liferay.oauth2.provider.model.OAuth2Application;
 import com.liferay.oauth2.provider.model.OAuth2RefreshToken;
 import com.liferay.oauth2.provider.model.OAuth2Token;
+import com.liferay.oauth2.provider.rest.tokenprovider.spi.BearerTokenProvider;
+import com.liferay.oauth2.provider.rest.tokenprovider.spi.BearerTokenProvider.AccessToken;
 import com.liferay.oauth2.provider.scopes.liferay.api.ScopeFinderLocator;
 import com.liferay.oauth2.provider.scopes.liferay.api.ScopedServiceTrackerMap;
-import com.liferay.oauth2.provider.scopes.spi.BearerTokenProvider;
-import com.liferay.oauth2.provider.scopes.spi.BearerTokenProvider.AccessToken;
 import com.liferay.oauth2.provider.service.OAuth2ApplicationLocalService;
 import com.liferay.oauth2.provider.service.OAuth2RefreshTokenLocalService;
 import com.liferay.oauth2.provider.service.OAuth2ScopeGrantLocalService;
